added comments and re-worked on code
